+++ app/views/homepage/index.html.erb
@@ -1,2 +1,2 @@
 <%= react_component("Hello", { name: @welcome_message }) %>
-<%= react_component("Tree", {values: [5, 3, 7, 1, 4, 11, 13]}) %>
+<%= react_component("TreeRenderer", {values: [1, 23, 4]}) %>
